+++ src/render/components/gain-selector/gain-selector.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input, Output, EventEmitter} from '@angular/core';
 import {IState} from '../../providers/state.service.';
-import {MAX_GAIN, MIN_GAIN} from '../../../common/constants';
+import {MAX_GAIN, MIN_GAIN, MIN_SAMPLE_RATE, MAX_SAMPLE_RATE} from '../../../common/constants';
 
 @Component({
     selector: 'gain-selector',
@@ -10,8 +10,11 @@ import {MAX_GAIN, MIN_GAIN} from '../../../common/constants';
 export class GainSelector {
     @Input() state: IState;
     @Output() setGain = new EventEmitter<number>();
+    @Output() setSampleRate = new EventEmitter<number>();
     @Output() toggleNormalization = new EventEmitter<boolean>();
     maxGain = MAX_GAIN;
     minGain = MIN_GAIN;
+    minSampleRate = MIN_SAMPLE_RATE;
+    maxSampleRate = MAX_SAMPLE_RATE;
 }